feat: coordinator checks withdrawal amounts against caps (#1526)
## Description

Closes https://github.com/stacks-network/sbtc/issues/1501.

## Changes

* Move the withdrawn total into the `SbtcLimits` struct. Maybe it
doesn't belong there, but it makes these changes much more simple.
* Update the `SbtcRequestsFilter` to filter withdrawal requests as well.
* Update the `SbtcRequestsFilter` to compute the minimum max-fee instead
of taking it as an input.
* Add an actual implementation to the in-memory version of the
`compute_withdrawn_total` function. This was so that one of the tests
would pass. I thought about just returning some dummy value, I can
revert to that.

## Testing Information

I'll be adding more test cases to this PR, but it's in good enough shape
to be reviewed.

## Checklist:

- [x] I have performed a self-review of my code