fix(tests): fixes for signer utxo tests (#1600)
## Description

These are all testing related changes.

This was done on the way to doing
https://github.com/stacks-sbtc/sbtc/issues/1598, but should probably be
merged regardless of what we do for
https://github.com/stacks-sbtc/sbtc/issues/1598, since it fixes one of
our tests.

## Changes

* Update how rotate keys transactions are stored and fetched in the
in-memory database. This change makes it possible to not use
`write_transaction` or `write_transactions`.
* Adds a new `TestBitcoinTxInfo` struct to simulate having a
`BitcoinTxInfo` struct, so that it can be deconstructed.
* Update the `TestData::push_bitcoin_txs` function to deconstruct the
given transactions into inputs and outputs. This change fixes the
`transaction_coordinator::tests::should_get_signer_utxo_unspent` test.

## Testing Information

This PR should only affect tests. It enables a previously ignored test
changes some tests, the signer UTXO ones, to go through the transaction
deconstruction code. This should more closely align with what happens in
production code.

## Checklist:

- [x] I have performed a self-review of my code