feat: fetch stacks block headers instead of the blocks (#1604)
## Description

Closes https://github.com/stacks-sbtc/sbtc/issues/1589.

This doesn't batch process stacks blocks, but it fetches block headers
so the signer is unlikely to OOM. Still, it does so all at once, writing
them to the database all at once as well.

After the conversion, headers take up 112 bytes each, assuming each
slimmed stacks block header has a bitcoin header hash and an 8 byte
block height. A signer can keep `17857142 = 2 * 1000 * 1000 * 1000 /
112` blocks with 2 GB of memory. Assuming 8000 blocks per day, that buys
us 6 years worth of stacks blocks.

## Changes

* Only take the headers for the nakamoto blocks fetched for each tenure.

## Testing Information

I still need to test this on devenv, but it is ready for review.


## Checklist:

- [x] I have performed a self-review of my code