fix: skip DKG if latest shares are unverified (#1705)
## Description

Currently if we need to DKG because of config changes we will keep doing
DKGs until we get a confirmed rotate key tx. But if we don't manage to
DKG + verify + rotate key submission in the same bitcoin tenure, we will
"discard" the previous progress and just start a new round.

## Changes

Prevent DKG if latest shares are unverified. 

## Testing Information

## Checklist:

- [x] I have performed a self-review of my code
- [ ] My changes generate no new warnings
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream
modules