fix: support brand new bootstrap signer set (#1752)
## Description

Closes https://github.com/stacks-sbtc/sbtc/issues/1751

## Changes

1. Load the aggregate key from the last DKG shares if we did not run DKG
and do not have any signer set info in the registry.
2. Signers proceed with verification using the latest verified DKG
shares if they do not have info in the registry. They fall back to the
latest shares if this is a contract deployment and the contracts have
not been deployed.

## Testing Information

Tested this on devenv with a fresh build and was able to sweep a deposit
without any issues.

Also, I redid one of our previously ignored tests and completely removed
two of the variants that were ignored. I couldn't get them to work and
got tired of attempting to fix them. Perhaps it would be better to keep
them around but ignored. I'm okay either way.

## Checklist:

- [x] I have performed a self-review of my code